⚡ Bolt: Centralize fast `esc` HTML escape function
Extracts the optimized single-pass `esc` function from `src/views/components.ts` into `src/shared/escape.ts` and replaces the unoptimized chained `.replace()` versions in `src/alerts/templates.ts` and `src/api/badge.ts`. This reduces GC pressure and speeds up string escaping across the whole application.
