+++ src/views/components.ts
@@ -23,39 +23,9 @@ const DMARC_TOOLTIPS: Record<string, string> = {
   ri: "Report interval — seconds between aggregate reports",
 };
 
-const HAS_ESCAPE_RE = /[&<>"']/;
+import { esc } from "../shared/escape.js";
 
-export function esc(s: string): string {
-  // ⚡ Bolt Optimization: Early return for strings that don't need escaping.
-  // Avoids unconditional regex replacements for the vast majority of strings.
-  // For strings that do need escaping, a single-pass loop avoids creating
-  // 5 intermediate strings, reducing GC pressure and making HTML rendering
-  // ~2-3x faster.
-  if (!HAS_ESCAPE_RE.test(s)) return s;
-
-  let res = "";
-  let last = 0;
-  for (let i = 0; i < s.length; i++) {
-    const c = s.charCodeAt(i);
-    if (c === 38) {
-      res += `${s.substring(last, i)}&amp;`;
-      last = i + 1;
-    } else if (c === 60) {
-      res += `${s.substring(last, i)}&lt;`;
-      last = i + 1;
-    } else if (c === 62) {
-      res += `${s.substring(last, i)}&gt;`;
-      last = i + 1;
-    } else if (c === 34) {
-      res += `${s.substring(last, i)}&quot;`;
-      last = i + 1;
-    } else if (c === 39) {
-      res += `${s.substring(last, i)}&#39;`;
-      last = i + 1;
-    }
-  }
-  return res + s.substring(last);
-}
+export { esc };
 
 export function gradeClass(grade: string): string {
   if (grade === "S") return "grade-s";
